--- a/promql/engine.go
+++ b/promql/engine.go
@@ -759,7 +759,7 @@ type EvalNodeHelper struct {
 	// For binary vector matching.
 	rightSigs    map[uint64]Sample
 	matchedSigs  map[uint64]map[uint64]struct{}
-	resultMetric map[uint64]labels.Labels
+	resultMetric map[[2]uint64]labels.Labels
 }
 
 // dropMetricName is a cached version of dropMetricName.
@@ -1614,14 +1614,11 @@ func signatureFunc(on bool, names ...string) func(labels.Labels) uint64 {
 // binary operation and the matching options.
 func resultMetric(lhs, rhs labels.Labels, op ItemType, matching *VectorMatching, enh *EvalNodeHelper) labels.Labels {
 	if enh.resultMetric == nil {
-		enh.resultMetric = make(map[uint64]labels.Labels, len(enh.out))
+		enh.resultMetric = make(map[[2]uint64]labels.Labels, len(enh.out))
 	}
 	// op and matching are always the same for a given node, so
 	// there's no need to include them in the hash key.
-	// If the lhs and rhs are the same then the xor would be 0,
-	// so add in one side to protect against that.
-	lh := lhs.Hash()
-	h := (lh ^ rhs.Hash()) + lh
+	h := [2]uint64{lhs.Hash(), rhs.Hash()}
 	if ret, ok := enh.resultMetric[h]; ok {
 		return ret
 	}