@roidelapluie 2020-01-04
```
benchmark                     old ns/op     new ns/op     delta
BenchmarkSignature-8          584           617           +5.65%
BenchmarkSignatureReuse-8     128           154           +20.31%

benchmark                     old allocs     new allocs     delta
BenchmarkSignature-8          7              7              +0.00%
BenchmarkSignatureReuse-8     1              1              +0.00%

benchmark                     old bytes     new bytes     delta
BenchmarkSignature-8          672           736           +9.52%
BenchmarkSignatureReuse-8     32            32            +0.00%
```

not good